Set default values for anchor and font
Using more recent version of graphics.py as reference, I tried to make the function in the older module
work more like the more recent module

References:
Code in Place
https://codeinplace.stanford.edu/cip3/forum?post=2da612db-5f8f-4516-b30c-9b953c3ace07
[The graphics library we're using was written for this course; it's not available on PyPI or through conda.  You can download it yourself if you want to look at the code, but it won't actually work offline -- it was designed to run in a browser.
However, there's a very similar graphics library that was written for CS106A, which does work offline.  There's documentation here, along with instructions for how to install it.  Note that this library isn't quite identical to the one we're using for Code in Place (for example, it's missing the create_polygon() function), but it's pretty close.]

https://codeinplace.stanford.edu/pyodide/graphics-0.0.4-py3-none-any.whl

CS106A | Spring
https://cs106a.stanford.edu/graphics

CS 106A
https://web.stanford.edu/class/archive/cs/cs106a/cs106a.1202/

Page not found · GitHub
https://github.com/nimaroger/cs106a-graphics-library/blob/master/graphics.py

graphics.py
http://nifty.stanford.edu/2013/denero-muralidharan-trends/graphics.py.html